PMA Complete

+++ driver.py
@@ -48,7 +48,7 @@ class Graphic():
             self.login()
         elif found == 1:
             messagebox.showinfo( "PMA", "Welcome Back User")
-            # self.home_page_user()
+            self.home_page_user()
         elif found == 2:
             messagebox.showinfo( "PMA", "Welcome Back Admin")
             self.home_page_admin()
@@ -89,7 +89,47 @@ class Graphic():
         self.root.quit()
         self.root.title("Product Market Analysis - Home Page")
         for i in self.components:
-            i.destroy()
+            if isinstance(i, list):
+                i.clear()
+            else:
+                i.destroy()
+
+
+        to_be_reviewed = []
+        admin =  open('Dataset/Login_Admin', 'r')
+        for line_num, line in enumerate(admin):
+            if "---->" in line:
+                item = str(line).split('####')
+                item = item[0].replace('---->','').replace("\n","")
+                to_be_reviewed.append(item)
+        admin.close()
+        
+        start = -1
+        already_reviewed = []
+        user =  open('Dataset/Login_User', 'r')
+        for line_num, line in enumerate(user):
+            if self.input_text==line:
+                start = line_num
+            if '_#_sep_#_' in line and line_num>start and start!=-1:
+                break
+            if '---->' in line and start!=-1 and line_num>start:
+                product = str(line).replace('---->','').replace('\n','')
+                already_reviewed.append(product)
+        user.close()
+
+        self.products = [x for x in to_be_reviewed if x not in already_reviewed]
+
+        self.var = IntVar()
+        self.value = 0
+
+        self.components, logo_icon = home_user(self.root.frame, self.products, self.username)
+        self.components[0].config(image=logo_icon)
+        self.components[3].config(command=self.logout)
+        for i in self.components[7]:
+            i.config(variable=self.var, value=self.value, command=self.review_item_user)
+            self.value+=1
+
+        self.root.mainloop()
 
     def home_page_admin(self):
         self.root.title("Product Market Analysis - Admin")
@@ -109,7 +149,7 @@ class Graphic():
             if '_#_sep_#_' in line and line_num>start and start!=-1:
                 break
             if '---->' in line and start!=-1 and line_num>start:
-                product = str(line).replace('---->','').replace('\n','')
+                product = str(line).replace('---->','').replace('\n','').replace("####","--")
                 products.append(product)
 
         self.var = IntVar()
@@ -117,7 +157,7 @@ class Graphic():
 
         self.components, logo_icon = home_admin(self.root.frame, products, self.username)
         self.components[0].config(image=logo_icon)
-        self.components[3].config(command=self.login)
+        self.components[3].config(command=self.logout)
         self.components[5].config(command=self.add_item_admin)
         for i in self.components[8]:
             i.config(variable=self.var, value=self.value, command=self.delete_item_admin)
@@ -127,11 +167,16 @@ class Graphic():
 
     def add_item_admin(self):
         top=self.top=Toplevel(self.root)
-        self.l=Label(top,text="Enter Item")
-        self.l.pack()
+        self.top.title("PMA")
+        self.top.config(bg="#0E0C24")
+        self.top.minsize(125, 125)
+        self.top.maxsize(125, 125)
+        self.top.geometry("125x125+100+100")
+        self.l=Label(top,text="Enter Item", fg="#0BFB2E",bg="#0E0C24", font="None 10")
+        self.l.pack(padx=(10, 10), pady=(10, 10))
         self.e=Entry(top)
-        self.e.pack()
-        self.b=Button(top,text='Ok',command=self.cleanup)
+        self.e.pack(padx=(10, 10), pady=(10, 10))
+        self.b=Button(top,text='Ok',command=self.cleanup, fg="#0BFB2E", bg="#0E0C24", bd=2, font="None 10")
         self.b.pack()
 
     def cleanup(self):
@@ -141,10 +186,29 @@ class Graphic():
         self.home_page_admin()
 
     def delete_item_admin(self):
-        choice = self.var.get()
-        delete_item(choice, self.input_text)
+        MsgBox = tk.messagebox.askquestion('Logout','Are you sure you want to delete this item?',icon = 'warning')
+        if MsgBox == 'yes':
+            choice = self.var.get()
+            delete_item(choice, self.input_text)
         self.home_page_admin()
         
+    def review_item_user(self):
+        choice = self.var.get()
+        item = self.products[choice]
+        MsgBox = tk.messagebox.askquestion('PMA','Are you sure you want to proceed with reviewing '+item, icon = 'warning')
+        if MsgBox == 'yes':
+            messagebox.showinfo( "PMA", "Reviewing will start now.\nPlease be honest about the product.\nThe system will capture your facial expressions to determine your experience with the product.\n\nPress 'q' to stop the reviewing")
+            self.root.quit()
+            average_emotion = start_emotion_prediction()
+            messagebox.showinfo( "PMA", "Review Recorded")
+            update_records(item, self.input_text, average_emotion)
+        self.home_page_user()
+
+    def logout(self):
+        MsgBox = tk.messagebox.askquestion('Logout','Are you sure you want to logout',icon = 'warning')
+        if MsgBox == 'yes':
+            self.login()
+
 def check_credentials(username, password):
     flag = 0
     input_text = username+"_#_sep_#_"+password+'\n'
@@ -192,6 +256,7 @@ def add_item(value, input_text):
 def delete_item(choice, input_text):
     lines = []
     start = -1
+    text = ''
     with open('Dataset/Login_Admin', 'r') as admin:
         for line_num, line in enumerate(admin):
             if input_text==line:
@@ -201,27 +266,58 @@ def delete_item(choice, input_text):
         lines = admin.readlines()
     admin.close()
     if start!=-1:
+        text = lines[start+1+choice]
         lines.remove(lines[start+1+choice])
     with open('Dataset/Login_Admin', 'w') as admin:
         admin.writelines(lines)
+    text = str(text).split('####')
+    with open('Dataset/Login_User', 'r') as user:
+        lines = user.readlines()
+    user.close()
+    text[0] = text[0]+"\n"
+    for i in lines:
+        if str(i)==str(text[0]):
+            lines.remove(i)
+    with open('Dataset/Login_User', 'w') as user:
+        user.writelines(lines)
 
+def update_records(item, input_text, average_emotion):
+    out_file = []
+    with open('Dataset/Login_User', 'r') as user:
+        in_file = user.readlines()
+        for line in in_file:
+            out_file.append(line)
+            if input_text==line:
+                out_file.append("---->"+str(item)+'\n')
+    user.close()
+    with open('Dataset/Login_User', 'w') as user:
+        user.writelines(out_file)
+
+    out_file = []
+    line_value = "---->"+str(item)
+    with open('Dataset/Login_Admin', 'r') as admin:
+        in_file = admin.readlines()
+        for line in in_file:
+            if str(line_value+'\n')==line:
+                out_file.append(line_value+"####"+average_emotion+'\n')
+            else:
+                out_file.append(line)
+    admin.close()
+    with open('Dataset/Login_Admin', 'w') as admin:
+        admin.writelines(out_file)
+
+def start_emotion_prediction():
+    try:
+        with open('Model/CNN_Model.json', 'r'):
+            model = load_model()
+            average_emotion = get_face(model)
+            return average_emotion
+    except FileNotFoundError:
+        create_cnn_model()
+        model = load_model()
+        average_emotion = get_face(model)
+        return average_emotion
     
-# def create_model():
-#     create_cnn_model()
-#     start_emotion_prediction()
-
-# def start_emotion_prediction():
-#     model = load_model()
-#     average_emotion = get_face(model)
-#     print(average_emotion)
-
 if __name__ == "__main__":
     os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2' 
-    # try:
-    #     with open('Model/CNN_Model.json', 'r'):
-    #         print("<-------------------Model Found------------------->")
-    #         start_emotion_prediction()
-    # except FileNotFoundError:
-    #     print("<-------------------Model does not exist. Creating one------------------->")
-    #     create_model()
     g = Graphic()
